+++ src/system-check/system-check.ts
@@ -34,7 +34,7 @@ export const getSystemCheckData = async (
     runVersionWithParse = runVersionAndParse,
 ): Promise<SystemCheckValues> => {
     const hasLlvmObjDumpOnPath = await checkLlvmObjDump();
-    const wperfVersion = await getWperfVersion(runVersionWithParse);
+    const wperfCommandAvailable = await hasWperfOnPath(runVersionWithParse);
     const hasWperfDriver = await getWperfDriver(platform, runDriverCheck);
 
     return {
@@ -48,7 +48,7 @@ export const getSystemCheckData = async (
             name: 'WPerf command available',
             description:
                 'This is required to run the commands used in this extension. Learn more https://gitlab.com/Linaro/WindowsPerf/windowsperf/-/blob/main/INSTALL.md or specify the absolute path to the WindowsPerf executable in the VSCode extension settings.',
-            isFound: !!wperfVersion,
+            isFound: wperfCommandAvailable,
         },
         isWperfDriverInstalled: {
             name: 'WPerf driver installed',
@@ -68,11 +68,18 @@ export const getSystemCheckData = async (
 export const getWperfVersion = async (
     runVersionWithParse = runVersionAndParse,
 ): Promise<string | undefined> => {
+    const versionJson = await runVersionWithParse();
+    return versionJson.find((a) => a.Component === 'wperf')?.Version;
+};
+
+export const hasWperfOnPath = async (
+    runVersionWithParse = runVersionAndParse,
+): Promise<boolean> => {
     try {
-        const versionJson = await runVersionWithParse();
-        return versionJson.find((a) => a.Component === 'wperf')?.Version;
+        await runVersionWithParse();
+        return true;
     } catch (error) {
-        return undefined;
+        return false;
     }
 };